// management/frontend/src/api/schemas/releaseVersionPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface ReleaseVersionPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 32
     */
  platform: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  version: string;
  /** @minimum 1 */
  build_number?: number;
  /** @maxLength 8000 */
  release_notes?: string;
  /** @maxLength 1000 */
  download_url?: string;
  /** @maxLength 32 */
  status?: string;
  is_current?: boolean;
  published_at?: string | null;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}